requirements.txt
# Core dependencies for the practicum environment.
# Install PyTorch separately using the official index for your CUDA version.

compressai==1.2.5
numpy>=1.23
scipy>=1.10
pillow>=9.5
torchmetrics>=1.3
tqdm>=4.66
pyyaml>=6.0
lpips>=0.1.4
scikit-image>=0.22
opencv-python>=4.8
tensorboard>=2.14
wandb>=0.16
einops>=0.7

